fix: fix work button reset
