+++ src/StorePage/filtres.ts
@@ -13,7 +13,7 @@ const divResetCopy = document.createElement("div");
 divResetCopy.classList.add("buttons-reset-copy");
 divFilters.appendChild(divResetCopy);
 
-const buttonReset = document.createElement("button");
+export const buttonReset = document.createElement("button");
 buttonReset.classList.add("filters__button");
 buttonReset.textContent = "Reset filters";
 divResetCopy.appendChild(buttonReset);
@@ -168,90 +168,4 @@ rangeContainerStock.appendChild(inputRangeMaxStock);
 document.querySelector("main")?.appendChild(fragmentStorePage);
 //
 
-const checkboxes: HTMLCollectionOf<Element> = document.getElementsByClassName("label-checkbox");
-const filter1 = new Set<string>();
-const filter2 = new Set<string>();
 
-Array.from(checkboxes).forEach((checkbox) => checkbox.addEventListener("change", (event) => {
-  //const cards: HTMLCollectionOf<Element> = document.getElementsByClassName("card-product");
-  const cards = document.querySelectorAll<HTMLElement>(".card-product");
-  const checkboxText: string = checkbox.textContent?.trim() as string;
-  const target = event.target as HTMLInputElement;
-
-  Array.from(cards).forEach((card) => {
-    const dataCategory = card.getAttribute("data-category") as string;
-    const dataBrand = card.getAttribute("data-brand") as string;
-
-    if ((checkboxText !== dataCategory)) {
-      card.classList.add("hide");
-      card.style.display = "none"
-      filter1.add(checkboxText);
-    }
-
-    if ((checkboxText !== dataBrand)) {
-      card.classList.add("hide");
-      card.style.display = "none"
-      filter1.add(checkboxText);
-    }
-
-    if (!target.checked) {
-      card.classList.remove("hide");
-      card.style.display = "flex"
-      filter1.delete(checkboxText);
-    }
-
-    if (filter1.has(dataCategory) && filter1.has(dataBrand)) {
-      card.classList.remove("hide");
-      card.style.display = "flex"
-      // eslint-disable-next-line no-dupe-else-if
-    } else if (filter1.has(dataCategory) || filter1.has(dataBrand)) {
-      card.classList.remove("hide");
-      card.style.display = "flex"
-    } else {
-      card.classList.add("hide");
-      card.style.display = "none"
-    }
-
-    if (card.classList.contains("hide") && filter1.size === 0) {
-      card.classList.remove("hide");
-      card.style.display = "flex";
-    }
-    const foundProduct = document.querySelector(".found-product");
-    found = Array.from(cards).filter(card => card.style.display === "flex").length;
-    if (foundProduct != undefined) foundProduct.textContent = `Found: ${found} pcs`;
-  });
-  console.log("filter1:", filter1);
-  console.log("filter2:", filter2);
-}));
-
-buttonReset.addEventListener("click", () => {
-  console.log("click");
-  const cards = document.querySelectorAll<HTMLElement>(".card-product");
-  Array.from(cards).forEach((card) => {
-    card.style.display = "flex";
-    filter1.clear();
-  });
-  location.reload();
-});
-
-document.querySelector(".input-search")?.addEventListener("input", (event) => {
-  const cards = document.querySelectorAll<HTMLElement>(".card-product");
-  const target = event.target as HTMLInputElement;
-
-  Array.from(cards).forEach((card) => {
-    const dataPrice = card.getAttribute("data-price") as string;
-    const dataRating = card.getAttribute("data-rating") as string;
-    const dataCategory = card.getAttribute("data-category") as string;
-    const dataBrand = card.getAttribute("data-brand") as string;
-
-    if (!(dataPrice.includes(target.value) || dataRating.includes(target.value) || dataCategory.includes(target.value.toLocaleUpperCase()) || dataBrand.includes(target.value.toLocaleUpperCase()))) {
-      card.style.display = "none"
-    } else {
-      card.style.display = "flex"
-    }
-  });
-  const foundProduct = document.querySelector(".found-product");
-  found = Array.from(cards).filter(card => card.style.display === "flex").length;
-  if (foundProduct != undefined) foundProduct.textContent = `Found: ${found} pcs`;
-});
-let found = 100;
